+++ src/Command.ts
@@ -4,6 +4,6 @@ export interface Command {
     name: string;
     description?: string;
     usage?: string;
-    action: (message: Message, args: string[]) => boolean | void;
+    action: (message: Message, args: string[]) => unknown;
     requiredArgCount?: number;
 }