+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### Next
+## 1.3.0
+#### Added
+- Promise based commands (async functions)
+
+### 1.2.0
 #### Added
 - Usage in command object for help message
-- Promise based commands (async functions)
 
 #### Changed
 - Command name can start with prefix or not. For example, with prefix `!`, command name can be `ping` or `!ping`